Organized review-related classes into review package.
