+++ core/src/main/java/com/aarves/bluepages/adapter/controllers/ReviewController.java
@@ -1,6 +1,6 @@
 package com.aarves.bluepages.adapter.controllers;
 
-import com.aarves.bluepages.usecase.interactors.ReviewManager;
+import com.aarves.bluepages.usecase.interactors.review.ReviewManager;
 
 public class ReviewController {
     private final ReviewManager reviewManager;
