+++ core/src/main/java/com/aarves/bluepages/adapter/presenters/ReviewPresenter.java
@@ -1,6 +1,6 @@
 package com.aarves.bluepages.adapter.presenters;
 
-import com.aarves.bluepages.usecase.interactors.ReviewOutputBoundary;
+import com.aarves.bluepages.usecase.interactors.review.ReviewOutputBoundary;
 
 public class ReviewPresenter implements ReviewOutputBoundary {
 }
