+++ core/src/main/java/com/aarves/bluepages/adapter/presenters/ReviewView.java
@@ -1,4 +1,7 @@
 package com.aarves.bluepages.adapter.presenters;
 
-public class ReviewView {
+public interface ReviewView {
+    void setReviews();
+
+    void finishActivity();
 }
