+++ core/src/main/java/com/aarves/bluepages/usecase/data/AccountDataAccess.java
@@ -1,7 +1,7 @@
 package com.aarves.bluepages.usecase.data;
 
 import com.aarves.bluepages.usecase.interactors.account.AccountDataBoundary;
-import com.aarves.bluepages.usecase.interactors.ReviewRepository;
+import com.aarves.bluepages.usecase.interactors.review.ReviewRepository;
 import com.aarves.bluepages.usecase.exceptions.PermissionsFailureException;
 import com.aarves.bluepages.entities.User;
