+++ core/src/main/java/com/aarves/bluepages/usecase/data/ReviewRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.aarves.bluepages.usecase.data;
 
-import com.aarves.bluepages.usecase.interactors.ReviewRepository;
+import com.aarves.bluepages.usecase.interactors.review.ReviewRepository;
 import com.aarves.bluepages.entities.Review;
 import com.aarves.bluepages.entities.User;
