--- core/src/main/java/com/aarves/bluepages/usecase/interactors/ReviewOutputBoundary.java
@@ -1,4 +1,0 @@
-package com.aarves.bluepages.usecase.interactors;
-
-public interface ReviewOutputBoundary {
-}
