+++ core/src/main/java/com/aarves/bluepages/usecase/interactors/review/ReviewManager.java
@@ -1,9 +1,10 @@
-package com.aarves.bluepages.usecase.interactors;
+package com.aarves.bluepages.usecase.interactors.review;
 
 import com.aarves.bluepages.usecase.exceptions.NotLoggedInException;
 import com.aarves.bluepages.usecase.exceptions.PermissionsFailureException;
 import com.aarves.bluepages.entities.Review;
 import com.aarves.bluepages.entities.User;
+import com.aarves.bluepages.usecase.interactors.Observer;
 
 import java.util.ArrayList;
 import java.util.List;
