+++ core/src/main/java/com/aarves/bluepages/usecase/interactors/review/ReviewOutputBoundary.java
@@ -1,0 +1,4 @@
+package com.aarves.bluepages.usecase.interactors.review;
+
+public interface ReviewOutputBoundary {
+}
